+++ CHANGELOG.md
@@ -9,6 +9,15 @@ The changelog format is based on [Keep a Changelog](https://keepachangelog.com/e
 * Moved CMake + conan + c++ package files and folders with cpp code inside src folder to be included in package
 
 ### Changed
+
+* Default agent_(input/output)_indexes is [] by default instead of None
+* Updated doc by running publish-interface-docs
+* Added feature to be able to initialize states using previously defined parameters or inputs
+  * This is done by setting "initializationVariable" = "{variable name}", instead of using the "name" and "start_value" attributes
+  * Added flag in schema to allow a variable to be reused when initializing states
+  * Allowed for multiple states to use the same variable for initialization
+  
+### Changed
 * Deleted azure files from old azure devops repo
 
 ### Changed